[IOTDB-2839] Add Python client CI (#5407)
