+++ client-py/requirements_dev.txt
@@ -19,7 +19,7 @@
 -r requirements.txt
 # Pytest to run tests
 pytest==6.2.2
-thrift==0.13.0
+thrift==0.14.1
 flake8==3.9.0
 black==20.8b1
 # For releases
