+++ client-py/tests/test_todf.py
@@ -65,7 +65,7 @@ def create_ts(session):
 
 
 def test_simple_query():
-    with IoTDBContainer() as db:
+    with IoTDBContainer("iotdb:dev") as db:
         db: IoTDBContainer
         session = Session(db.get_container_host_ip(), db.get_exposed_port(6667))
         session.open(False)
@@ -92,7 +92,7 @@ def test_simple_query():
 
         df_input.insert(0, "Time", timestamps)
 
-        session_data_set = session.execute_query_statement("SELECT * FROM root.*")
+        session_data_set = session.execute_query_statement("SELECT ** FROM root")
         df_output = session_data_set.todf()
         df_output = df_output[df_input.columns.tolist()]
 
@@ -101,7 +101,7 @@ def test_simple_query():
 
 
 def test_with_null_query():
-    with IoTDBContainer() as db:
+    with IoTDBContainer("iotdb:dev") as db:
         db: IoTDBContainer
         session = Session(db.get_container_host_ip(), db.get_exposed_port(6667))
         session.open(False)
@@ -171,7 +171,7 @@ def test_with_null_query():
 
         df_input.insert(0, "Time", timestamps)
 
-        session_data_set = session.execute_query_statement("SELECT * FROM root.*")
+        session_data_set = session.execute_query_statement("SELECT ** FROM root")
         df_output = session_data_set.todf()
         df_output = df_output[df_input.columns.tolist()]
 
@@ -180,7 +180,7 @@ def test_with_null_query():
 
 
 def test_multi_fetch():
-    with IoTDBContainer() as db:
+    with IoTDBContainer("iotdb:dev") as db:
         db: IoTDBContainer
         session = Session(db.get_container_host_ip(), db.get_exposed_port(6667))
         session.open(False)
@@ -207,7 +207,7 @@ def test_multi_fetch():
 
         df_input.insert(0, "Time", timestamps)
 
-        session_data_set = session.execute_query_statement("SELECT * FROM root.*")
+        session_data_set = session.execute_query_statement("SELECT ** FROM root")
         session_data_set.set_fetch_size(100)
         df_output = session_data_set.todf()
         df_output = df_output[df_input.columns.tolist()]
